requirements.txt
slack-bolt==1.16.1
python-dotenv==1.0.0
SQLAlchemy==1.4.46
apscheduler==3.10.1
pytest==7.3.1
